+++ dsp-assignment-2-mapreduce/src/main/java/dsp/IntSumReducer.java
@@ -1,21 +1,21 @@
 package dsp;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
 public class IntSumReducer
-		extends Reducer<Text, IntWritable, Text, IntWritable> {
-	private IntWritable result = new IntWritable();
+		extends Reducer<Text, LongWritable, Text, LongWritable> {
+	private LongWritable result = new LongWritable();
 
 	@Override
-	public void reduce(Text key, Iterable<IntWritable> values,
+	public void reduce(Text key, Iterable<LongWritable> values,
 					   Context context
 	) throws IOException, InterruptedException {
-		int sum = 0;
-		for (IntWritable val : values) {
+		long sum = 0;
+		for (LongWritable val : values) {
 			sum += val.get();
 		}
 		result.set(sum);
