+++ dsp-assignment-2-mapreduce/src/test/java/dsp/TestMapReduce.java
@@ -1,6 +1,6 @@
 package dsp;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.mrunit.mapreduce.MapReduceDriver;
@@ -11,13 +11,13 @@ import java.io.IOException;
 
 public class TestMapReduce {
 
-	private MapReduceDriver<Object, Text, Text, IntWritable, Text, IntWritable> mapReduceDriver;
+	private MapReduceDriver<Object, Text, Text, LongWritable, Text, LongWritable> mapReduceDriver;
 
 	@Before
 	public void setUp() {
 		TokenizerMapper mapper = new TokenizerMapper();
 
-		Reducer<Text, IntWritable, Text, IntWritable> reducer = new IntSumReducer();
+		Reducer<Text, LongWritable, Text, LongWritable> reducer = new IntSumReducer();
 
 		mapReduceDriver = MapReduceDriver.newMapReduceDriver(mapper, reducer);
 	}
@@ -27,14 +27,14 @@ public class TestMapReduce {
 		addInput("' leaves open the possibility\t1981\t4\t4\t3");
 		addInput("\"\"\" During the final stages\"\t1994\t3\t3\t3\n");
 		addInput("\"! \"\" \"\" It ought\"\t1893\t1\t1\t1");
-		mapReduceDriver.withOutput(new Text("1980\tleaves"), new IntWritable(4));
-		mapReduceDriver.withOutput(new Text("1980\tleaves open"), new IntWritable(4));
-		mapReduceDriver.withOutput(new Text("1980\topen"), new IntWritable(4));
-		mapReduceDriver.withOutput(new Text("1980\topen possibility"), new IntWritable(4));
-		mapReduceDriver.withOutput(new Text("1980\tpossibility"), new IntWritable(4));
-		mapReduceDriver.withOutput(new Text("1990\tfinal"), new IntWritable(3));
-		mapReduceDriver.withOutput(new Text("1990\tfinal stages"), new IntWritable(3));
-		mapReduceDriver.withOutput(new Text("1990\tstages"), new IntWritable(3));
+		mapReduceDriver.withOutput(new Text("1980\tleaves"), new LongWritable(4));
+		mapReduceDriver.withOutput(new Text("1980\tleaves open"), new LongWritable(4));
+		mapReduceDriver.withOutput(new Text("1980\topen"), new LongWritable(4));
+		mapReduceDriver.withOutput(new Text("1980\topen possibility"), new LongWritable(4));
+		mapReduceDriver.withOutput(new Text("1980\tpossibility"), new LongWritable(4));
+		mapReduceDriver.withOutput(new Text("1990\tfinal"), new LongWritable(3));
+		mapReduceDriver.withOutput(new Text("1990\tfinal stages"), new LongWritable(3));
+		mapReduceDriver.withOutput(new Text("1990\tstages"), new LongWritable(3));
 
 		mapReduceDriver.runTest();
 	}
@@ -45,9 +45,9 @@ public class TestMapReduce {
 	@Test
 	public void testMapReduceRegression() throws IOException {
 		addInput("\"\"\" 2 Cor . iv\"\t1900\t2\t2\t2");
-		mapReduceDriver.withOutput(new Text("1900\tcor"), new IntWritable(2));
-		mapReduceDriver.withOutput(new Text("1900\tcor iv"), new IntWritable(2));
-		mapReduceDriver.withOutput(new Text("1900\tiv"), new IntWritable(2));
+		mapReduceDriver.withOutput(new Text("1900\tcor"), new LongWritable(2));
+		mapReduceDriver.withOutput(new Text("1900\tcor iv"), new LongWritable(2));
+		mapReduceDriver.withOutput(new Text("1900\tiv"), new LongWritable(2));
 		mapReduceDriver.runTest();
 	}
